+++ jgclark.Dashboard/CHANGELOG.md
@@ -2,9 +2,16 @@
 For more details see the [plugin's documentation](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
 
 ## [2.1.0.b4] 2024-12-???
+### New
 - New "Apply to sections under headings in Calendar notes?" setting. When turned on, then the phrases in the previous setting will be used to ignore whole sections in Calendar notes, if that section's heading contains any of those phrases.
+### Changed
 - Turn off 'Referenced' Calendar sections if they have no items to display.
 - Optimised Header layout to work on iPhones, iPad as well as macOS sizes. (On iPhone the Refresh button and info is turned off completely.)
+- Brought display of dropdown menus into line with other similar input fields.
+### Fixed
+- Fixed edge case where '>This Week' section could be shown twice.
+### Dev only changes
+- (dbw) Changed rollup to prefix most Dashboard CSS imports with .dashboard.
 
 ## [2.1.0.b3] 2024-12-18
 ### New